Time: 61 ms (38.59%), Space: 98.6 MB (17.19%) - LeetHub

+++ 2187-minimum-time-to-complete-trips/2187-minimum-time-to-complete-trips.cpp
@@ -6,13 +6,13 @@ public:
         for(int i=0;i<n;i++){
             tt += mid/(long long)time[i];
         }
-        if(tt < totalTrips) return false;
+        if(tt < (long long)totalTrips) return false;
         return true;
     }
     long long minimumTime(vector<int>& time, int totalTrips) {
         int n = time.size();
         long long max = time[0];
-        for(int t : time){
+        for(long long t : time){
             if(max < t) max = t;
         }
         long long lo = 1;
